#12 #13 Create tab bar, create page to list questionnaires
I have implemented a tab bar with 3 items; the welcome screen, the
questionnaires screen and the 6 minute walk test screen. I have also
created the page which will list the questionnaires available for the
patient to answer. As of current, there is only placeholder text, but
the real questionnaires will be implemented later.