#3 #4 #5 Create questionnaire pages
I have created the base questionnaire page which appears when a user
selects a questionnaire. From this page, patients can choose an answer,
and click "next" to proceed to the next question, until all questions
are answered. I am yet to implement the actual questions, and as of
current there are only placeholder questions.