#3 #4 #5 Include slider for questions which require it
Previously, the only question type which could be displayed was multiple
choice questions. Looking through the questionnaires, some require a
scale (e.g. from 0 to 100). To represent this on the app, I have
implemented a question of type "slider" which shows a slider in which
the patient can select a range from 0 to 100.