#3 #4 #5 #13 Change button style and change navigation items
I have changed the style of the "begin" and "next" buttons. They now
have white text and a blue background. I have also disabled the buttons.
They are reenabled when I user selects an item from the table view. On
top of this, I have changed the top left navigation item. Upon selecting
a questionnaire, the "back" button previously displayed "Questionnaires"
which pushed the HRQoL title to the side. I have changed this to "HRQoL"
and the title seems to be centered as it should be.