#8 #17 Add audio to walk test
I have added audio queues to the walk test. As the rate at which the
pedometer cannot be set, I have included a beginning audio track and an
ending audio track. This is to make sure that the step count doesn't
include steps made from a few seconds before, and so the final step
count includes steps made from a few seconds before.

Closes #17